opea-comps/megaservice-vqa/services/vqa/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
Pillow==9.5.0
transformers==4.30.2
torch>=2.0.0
accelerate>=0.20.3
opencv-python-headless==4.7.0.72
python-multipart==0.0.6